+++ rift_egui/src/components/signature_information.rs
@@ -11,11 +11,11 @@ pub fn show_signature_information(
         x: (state.relative_cursor.column as f32 * char_width)
             + top_left.x
             + char_width
-            + state.preferences.editor_padding,
+            + state.preferences.editor_padding as f32,
         y: ((state.relative_cursor.row - 1) as f32 * char_height)
             + top_left.y
             + char_height
-            + state.preferences.editor_padding,
+            + state.preferences.editor_padding as f32,
     };
     egui::Window::new("signature_information")
         .movable(false)
